requirements.txt
openai
sounddevice
scipy
python-dotenv